Terminado el builder de mediciones, creados y agregados metodos a los repos, agregados un monton de excepciones del MedicionBuilder. Faltan test del package Medicion y de las excepciones, lo demás ni idea.

+++ src/main/java/transporte/PropulsionHumana.java
@@ -15,7 +15,7 @@ public class PropulsionHumana implements Transporte {
 
 
   @Override
-  public int distanciaEntre(PuntoUbicacion origen, PuntoUbicacion destino) throws IOException {
+  public double distanciaEntre(PuntoUbicacion origen, PuntoUbicacion destino) throws IOException {
 
     ServicioGeodds servicioGeodds = ServicioGeodds.getInstancia();
     Distancia distancia = servicioGeodds
