+++ src/main/java/transporte/Transporte.java
@@ -4,7 +4,7 @@ import java.io.IOException;
 import linea.PuntoUbicacion;
 
 public interface Transporte {
-  public int distanciaEntre(PuntoUbicacion origen, PuntoUbicacion destino) throws IOException;
+  public double distanciaEntre(PuntoUbicacion origen, PuntoUbicacion destino) throws IOException;
 }
