+++ src/main/java/transporte/VehiculoParticular.java
@@ -25,7 +25,7 @@ public class VehiculoParticular implements Transporte {
   }
 
   @Override
-  public int distanciaEntre(PuntoUbicacion origen, PuntoUbicacion destino) throws IOException {
+  public double distanciaEntre(PuntoUbicacion origen, PuntoUbicacion destino) throws IOException {
 
     ServicioGeodds servicioGeodds = ServicioGeodds.getInstancia();
     Distancia distancia = servicioGeodds.distancia(
